Check if valid FMan handle exists on FM_PCD_Open and FM_PORT_Open
This patch checks if a valid FMan handle exists when invoking FM_PCD_Open and FM_PORT_Open in a particular context: call fmc tool by a user that does not have  privileges to open the FMan device. In this case a segmentation fault will be generated because the FMan device will not be opened and all the operations on the PCD will be invalid. The patch fixes the above path.

Signed-off-by: Florinel Iordache <florinel.iordache@nxp.com>