fm_lib: make REPORT_ERROR() definition more function-like
gcc warns:

src/fm_lib.c: In function ‘FM_ForceIntr’:
src/fm_lib.c:297:57: warning: suggest braces around empty body in an ‘if’ statement [-Wempty-body]
  297 |         REPORT_ERROR(MINOR, E_INVALID_OPERATION, NO_MSG);
      |                                                         ^
src/fm_lib.c: In function ‘FM_GetApiVersion’:
src/fm_lib.c:313:57: warning: suggest braces around empty body in an ‘if’ statement [-Wempty-body]
  313 |         REPORT_ERROR(MINOR, E_INVALID_OPERATION, NO_MSG);
      |

This happens when we compile with DEBUG_ERRORS undefined.

The full code is:

    if (ioctl(p_Dev->fd, FM_IOC_FORCE_INTR, (ioc_fm_exceptions) exception))
        REPORT_ERROR(MINOR, E_INVALID_OPERATION, NO_MSG);

    if (ioctl(p_Dev->fd, FM_IOC_GET_API_VERSION, p_version))
        REPORT_ERROR(MINOR, E_INVALID_OPERATION, NO_MSG);

Because the REPORT_ERROR() macro evaluates to an empty expression in
this case, we are left with just ";" for the "if" body. This is fine
functionally, but -Wempty-body wants it to be more explicit.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>